Remove puts/gets from classes && Create 1 file for each class
